Update create_icons.sh
Updated comments.

Added important note regarding how the script is run. If it is run as sudo, it WON'T update the launcher settings due to the set function of gsettings only updating the settings for the user running the command. Therefore it needs to be run as the user whose settings need modified. Going to have to think about this approach further.