Update install_desktop_icons.sh
Added counts to display the number of files modified/created.
Changed variable names to be more clear.
Added a couple more checks on the icon operation to make it much more strict. In other words, the directory MUST now exist and it MUST have something in it for the cp to be attempted.