Update create_launcher_icons.sh
Added extra comments.
Added a check to make sure the launcher_desktop directory exists where it should exist. If it doesn't, the program errors out.
Fixed bug with the localLauncherDir not working when calling the script at login. With the full path, it works as expected. Although, this means this must be changed when changing the directory of the script.