Create create_vbvm.sh
Added the core logic. Some additional work needs to be done for the comment parsing (in-line comments for instance). Also need to make sure that when unregistering the VM (in the event of an error) is the correct move. Alternatively, that command can be moved to the end of the program so that if any error occurs in reading the file or storing the data, then it won't be created.