Update create_vbvm.sh
Added an option for the a share. Note, this will not be accurate if the path has a space in it. I plan to investigate this further in a future revision. I also plan to add the ability for n-number of shared folders to be added similar to how I did the NICs.